requirements.txt
beautifulsoup4
filetype
httpx
latest-user-agents
progressbar2
